Fix wrong language syntax detection in test
When running`make test`, the test creates a file which will be detected as a valid bash script. As the pre-commit script attempts to detect the language used in the file, it runs a small test for each language one by one, and the test for detecting bash scripts will always return true. If the detection reaches this point, the script will not go any further and files in languages such as PHP, XML, and YAML will all be linted as bash scripts.

To solve the issue, the test for detecting bash scripts has been moved to the end of the list and will only be run if all other language detection return false.